backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

pyotp==2.9.0
qrcode==7.4.2
python-jose==3.3.0

pydantic>=2.12,<3
